[WIP] Seperate profile matching from instrospection
Currently, the edeploy plugin to ironic-discoverd is doing the profile
matching, but it only deals with data posted by the discovery ramdisk.
Running the benchmarks can be very time consuming on a typical
production server, and we already store the results in the ironic db.
The benchmarks should not vary much between runs, so this time is wasted
in future runs.

This patch provides a utility to run matching seperate from
introspection.

Change-Id: I2b37cbfbf49019c655b6d95e3a4f343c45995881