Consume edeploy facts from Swift
The edeploy plugin in ironic-discoverd collects too many
facts to store them in the Ironic database. It has changed
to store them in a Swift object instead. As such,
ahc-tools needs to get the facts from Swift.

Since we now have potentially two sets of credentials(
Ironic, Swift), oslo config is used for both ahc-match
and ahc-report, instead of relying on environment
variables for the Ironic credentials.

Change-Id: I38681a7c10661ae498299fcc79eac75b8fb65ef0