Only match on manageable or available nodes
Once a node is associated to a nova instance, it no longer
makes sense to include it in matching or reporting. It
also does not make sense to match failed nodes.

Unfortunately, there is not currently a filter in the
Ironic API for provision_state. For now, we will iterate
over the nodes and check the provision_state. Added a
"FIXME" to change this to a filter once it is implemented
in Ironic.

Change-Id: I05cae7938d8059f0ce38b94d3a12dd9a02dfae99